+++ src/coreclr/vm/assemblyspec.hpp
@@ -16,7 +16,7 @@
 #define _ASSEMBLYSPEC_H
 #include "hash.h"
 #include "assemblyspecbase.h"
-#include "domainfile.h"
+#include "domainassembly.h"
 #include "holder.h"
 
 class AppDomain;
