+++ src/coreclr/vm/coreassemblyspec.cpp
@@ -15,7 +15,7 @@
 #include "appdomain.inl"
 #include <peimage.h>
 #include "peimagelayout.inl"
-#include "domainfile.h"
+#include "domainassembly.h"
 #include "holder.h"
 #include "bundle.h"
 #include "strongnameinternal.h"
